[x86_64] Clean up multitasking code

+++ kernel/kernel/drivers/pit/pit.h
@@ -19,6 +19,6 @@ void pit_timer_init(uint32_t frequency);
 uint32_t pit_clock();
 uint32_t tick_count();
 
-uint32_t ms_since_boot(void);
+uintptr_t ms_since_boot(void);
 
 #endif
